Announcement Deletion
implement announcement deletion

+++ app/Http/Controllers/AnnouncementController.php
@@ -32,15 +32,16 @@ class AnnouncementController extends Controller
 
     public function destroy($id)
     {
-        $announcement = DB::select('select * from announcements where id = ?', [$id]);
+        $announcement = DB::select('select * from announcements where id = ? limit 1', [$id]);
         $creator = AnnouncementCreator::where('annc_id', $id)->where('u_id', Auth::id())->first();
 
         if(!$creator){
             return redirect()->route('announcements')->with('error', 'You are not the announcement creator! User not authorized to delete this announcement!');
         }
 
-        $creator->delete();
-        $announcement->delete();
+        DB::transaction(function () use ($id, $announcement, $creator) {DB::delete('delete from announcement_creator where annc_id = ?', [$id]);});
+        DB::delete('delete from announcements where id = ?', [$id]);
+
         return redirect()->route('announcements')->with('success', 'Announcement deleted successfully!');
     }
 }
